feat: tick counter now keeps increasing

+++ network/src/lib.rs
@@ -8,7 +8,7 @@ pub enum ServerPacket {
         no_of_chunks: u32,
         chunks_per_player: u16,
         snapshots: Vec<ChunkSnapshot>,
-        tick_count: u8,
+        tick_count: u64,
     },
     Snapshot(ChunkSnapshot),
     Delta(ChunkDelta),
